+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/TxBody.hs
@@ -1066,7 +1066,7 @@ dijkstraRedeemerPointer ::
   StrictMaybe (DijkstraPlutusPurpose AsIx era)
 dijkstraRedeemerPointer txBody = \case
   DijkstraMinting policyID ->
-    DijkstraMinting <$> indexOf policyID (txBody ^. mintPoliciesTxBodyF)
+    DijkstraMinting <$> indexOf policyID (txBody ^. forgingPoliciesTxBodyF)
   DijkstraSpending txIn ->
     DijkstraSpending <$> indexOf txIn (txBody ^. inputsTxBodyL)
   DijkstraWithdrawing accountAddress ->
@@ -1088,7 +1088,7 @@ dijkstraRedeemerPointerInverse ::
   StrictMaybe (DijkstraPlutusPurpose AsIxItem era)
 dijkstraRedeemerPointerInverse txBody = \case
   DijkstraMinting idx ->
-    DijkstraMinting <$> fromIndex idx (txBody ^. mintPoliciesTxBodyF)
+    DijkstraMinting <$> fromIndex idx (txBody ^. forgingPoliciesTxBodyF)
   DijkstraSpending idx ->
     DijkstraSpending <$> fromIndex idx (txBody ^. inputsTxBodyL)
   DijkstraWithdrawing idx ->
@@ -1149,8 +1149,11 @@ instance
   ) =>
   MaryEraTxBody DijkstraEra
   where
-  mintTxBodyL = memoRawTypeL @DijkstraEra . mintDijkstraTxBodyRawL
-  {-# INLINE mintTxBodyL #-}
+  forgingTxBodyL =
+    memoRawTypeL @DijkstraEra
+      . mintDijkstraTxBodyRawL
+      . lens Forging (\_ -> unForging)
+  {-# INLINE forgingTxBodyL #-}
 
 collateralInputsDijkstraTxBodyRawL :: Lens' (DijkstraTxBodyRaw TopTx era) (Set TxIn)
 collateralInputsDijkstraTxBodyRawL =