+++ eras/mary/impl/src/Cardano/Ledger/Mary/Core.hs
@@ -1,9 +1,8 @@
 module Cardano.Ledger.Mary.Core (
   MaryEraTxBody (..),
-  forgingTxBodyL,
   mintedAssetsTxBodyF,
   burnedAssetsTxBodyF,
-  mintPoliciesTxBodyF,
+  forgingPoliciesTxBodyF,
   module Cardano.Ledger.Mary.Forging,
   module Cardano.Ledger.Allegra.Core,
 ) where
@@ -14,7 +13,6 @@ import Cardano.Ledger.Mary.Tx ()
 import Cardano.Ledger.Mary.TxBody (
   MaryEraTxBody (..),
   burnedAssetsTxBodyF,
-  forgingTxBodyL,
-  mintPoliciesTxBodyF,
+  forgingPoliciesTxBodyF,
   mintedAssetsTxBodyF,
  )