+++ eras/mary/impl/src/Cardano/Ledger/Mary/UTxO.hs
@@ -107,4 +107,4 @@ getMaryScriptsNeeded u txBody =
   case getShelleyScriptsNeeded u txBody of
     ShelleyScriptsNeeded shelleyScriptsNeeded ->
       ShelleyScriptsNeeded $
-        shelleyScriptsNeeded `Set.union` Set.map policyID (txBody ^. mintPoliciesTxBodyF)
+        shelleyScriptsNeeded `Set.union` Set.map policyID (txBody ^. forgingPoliciesTxBodyF)